refactor(scheduler): move backfill logic into backfill.go
Extract handleBackfill, effectiveFireOverlap, countCronFires, and
processBackfills from workflow.go into a dedicated backfill.go. All four
functions are specific to the backfill signal path and have no callers
outside that path. workflow.go retains the call site and the shared fire
dispatch helpers (processScheduleFire, tryStartFire, drainBufferedFires)
that are used by both backfill and the live cron timer path.

No behavior change.
